backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pydantic[email]==2.6.4
python-dotenv==1.0.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
httpx==0.27.0